% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{ggmap_bbox}
\alias{ggmap_bbox}
\title{Hacking ggmap bbox to work with sf}
\usage{
ggmap_bbox(map)
}
\arguments{
\item{map}{A ggamp object}
}
\value{
a ggmap object with the bounding box in 3857
}
\description{
Hacking ggmap bbox to work with sf
}
