% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{tw_shp_crop}
\alias{tw_shp_crop}
\title{臺灣本島cropped out simple feature}
\format{
A data frame with 22 rows and 5 columns.
}
\description{
A detailed description of what the data frame contains and any other necessary details.
}
