% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{tw_shp_crop_election2020}
\alias{tw_shp_crop_election2020}
\title{2020臺灣總統大選投票資料}
\format{
A data frame with 66 rows and 18 columns.
}
\description{
A detailed description of what the data frame contains and any other necessary details.
}
